avril/requirements/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
coreapi==2.0.4
Django==1.8.14
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.15.0
django-guardian==1.4.6
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
itypes==1.1.0
kombu==3.0.37
Markdown==2.6.6
Pygments==2.1.3
PyJWT==1.4.2
PyMySQL==0.7.9
pytz==2016.7
requests==2.11.1
six==1.10.0
uritemplate==3.0.0
